bedrock2: hardcode cases for access_size

+++ bedrock2/src/Examples/Demos.v
@@ -123,7 +123,7 @@ Section Demos.
   Definition bsearch: Prog := ("bsearch", ([left; right; target], [left], bedrock_func_body:(
     while (left < right) {{
       mid = left + (((right-left) >> 4) << 3);;
-      tmp = *(uint64_t*) mid;;
+      tmp = *(uintptr_t*) mid;;
       if (tmp < target) {{
         left = mid + 8
       }} else {{
